cleanup/add report delete button
